2026/day-34/requirements.txt
flask
redis
mysql-connector-python
